+++ inst/sql/oracle/test.sql
@@ -1,1 +1,3 @@
-SELECT a FROM table;
+{DEFAULT @some_id = 123}
+
+SELECT a FROM my_table WHERE my_id = @some_id;